# Conflicts:
#	frontend/web/css/admin/styles.css
#	frontend/web/css/styles.css
#	frontend/web/css/styles.min.css
#	frontend/web/js/admin/scripts.js
#	frontend/web/js/scripts.js

+++ frontend/modules/actions/controllers/AdminController.php
@@ -7,6 +7,7 @@ use frontend\modules\actions\models\Actions;
 use frontend\modules\actions\models\ActionsSearch;
 use frontend\modules\actions\models\ActionsConditions;
 use frontend\modules\actions\models\ActionsActions;
+use frontend\modules\meta\models\Meta;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
@@ -34,9 +35,6 @@ class AdminController extends Controller
 
     public function beforeAction($action)
     {
-        if (!Yii::$app->user->identity->is_admin) {
-            throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
-        }
         $this->layout = '@app/views/layouts/admin.twig';
         return true;
     }
@@ -47,12 +45,29 @@ class AdminController extends Controller
      */
     public function actionIndex()
     {
+        if (Yii::$app->user->isGuest || !Yii::$app->user->can('ActionsView')) {
+            throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
+            return false;
+        }
+
         $searchModel = new ActionsSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
         return $this->render('index.twig', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
+            'tableData' => [
+                'joined' => function ($model) {
+                    return $model['joined'] > 0 ?
+                        '<a href="/admin/users?joined_to='.$model['uid'].'">'.$model['joined'].'</a>' :
+                        $model['joined'];
+                },
+                'completed' => function ($model) {
+                    return $model['completed'] > 0 ?
+                        '<a href="/admin/users?completed_to='.$model['uid'].'">'.$model['completed'].'</a>' :
+                        $model['completed'];
+                },
+            ],
         ]);
     }
 
@@ -75,8 +90,17 @@ class AdminController extends Controller
      */
     public function actionCreate()
     {
+        if (Yii::$app->user->isGuest || !Yii::$app->user->can('ActionsCreate')) {
+            throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
+            return false;
+        }
+
         $model = new Actions();
         $promos = ArrayHelper::map(DbPromo::find()->all(), 'uid', 'title');
+        $pages = ArrayHelper::map(Meta::find()
+            ->select(['page'])
+            ->where(['not like', 'page', '*'])
+            ->all(), 'page', 'page');
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['index']);
@@ -84,6 +108,7 @@ class AdminController extends Controller
             return $this->render('create.twig', [
                 'model' => $model,
                 'promos' => $promos,
+                'pages' => $pages,
             ]);
         }
     }
@@ -96,8 +121,17 @@ class AdminController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (Yii::$app->user->isGuest || !Yii::$app->user->can('ActionsEdit')) {
+            throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
+            return false;
+        }
+
         $model = $this->findModel($id);
         $promos = ArrayHelper::map(DbPromo::find()->all(), 'uid', 'title');
+        $pages = ArrayHelper::map(Meta::find()
+            ->select(['page'])
+            ->where(['not like', 'page', '*'])
+            ->all(), 'page', 'page');
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['index']);
@@ -105,6 +139,7 @@ class AdminController extends Controller
             return $this->render('update.twig', [
                 'model' => $model,
                 'promos' => $promos,
+                'pages' => $pages,
             ]);
         }
     }
@@ -117,6 +152,11 @@ class AdminController extends Controller
      */
     public function actionDelete($id)
     {
+        if (Yii::$app->user->isGuest || !Yii::$app->user->can('ActionsDelete')) {
+            throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
+            return false;
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
@@ -128,6 +168,11 @@ class AdminController extends Controller
      */
     public function actionAddCondition($id)
     {
+        if (Yii::$app->user->isGuest || !Yii::$app->user->can('ActionsEdit')) {
+            throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
+            return false;
+        }
+
         $model = new ActionsConditions();
         $model->action_id = $id;
 
@@ -160,8 +205,12 @@ class AdminController extends Controller
      */
     public function actionUpdateCondition($id)
     {
-        $model = $this->findCondition($id);
+        if (Yii::$app->user->isGuest || !Yii::$app->user->can('ActionsEdit')) {
+            throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
+            return false;
+        }
 
+        $model = $this->findCondition($id);
 
         $dictionary = Yii::$app->params['dictionary'];
         $loyaltyStatuses = isset($dictionary['loyalty_status']) ? $dictionary['loyalty_status'] : [];
@@ -193,6 +242,11 @@ class AdminController extends Controller
      */
     public function actionDeleteCondition($id)
     {
+        if (Yii::$app->user->isGuest || !Yii::$app->user->can('ActionsEdit')) {
+            throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
+            return false;
+        }
+
         $model = $this->findCondition($id);
         $actionId = $model->action_id;
         $model->delete();
@@ -207,6 +261,11 @@ class AdminController extends Controller
      */
     public function actionAddAction($id)
     {
+        if (Yii::$app->user->isGuest || !Yii::$app->user->can('ActionsEdit')) {
+            throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
+            return false;
+        }
+
         $model = new ActionsActions();
         $model->action_id = $id;
 
@@ -227,6 +286,11 @@ class AdminController extends Controller
      */
     public function actionUpdateAction($id)
     {
+        if (Yii::$app->user->isGuest || !Yii::$app->user->can('ActionsEdit')) {
+            throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
+            return false;
+        }
+
         $model = $this->findAction($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -248,6 +312,11 @@ class AdminController extends Controller
      */
     public function actionDeleteAction($id)
     {
+        if (Yii::$app->user->isGuest || !Yii::$app->user->can('ActionsEdit')) {
+            throw new \yii\web\ForbiddenHttpException('Просмотр данной страницы запрещен.');
+            return false;
+        }
+
         $model = $this->findAction($id);
         $actionId = $model->action_id;
         $model->delete();
